remove "sillier" section
It felt like the sillier second part of this article detracted from the
main point, so I've removed it. It was mostly about the topic of scoping
intermediate variables in conditional statements, and that can be a
separate article.

Additionally, I added an actual return statemtent to the last example to
keep it consistent with all the other examples.