don't shame iota, just don't bring it up
iota is not the right tool here. It has a specific set of things it's
good for, but it causes more headache than it helps in the scenarios
described in this article. I don't want to see it used, but there's no
reason I should shame it's usage, therefor I just won't mention it.